Consolidate frontend code under one project and use folder by feature
