+++ CodeHub.sln
@@ -22,8 +22,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeHub.Api", "src\backend\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Frontend", "Frontend", "{FD7D4EDA-3EEC-4989-B479-563417E6848F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeHub.Portal.Services", "src\frontend\CodeHub.Portal.Services\CodeHub.Portal.Services.csproj", "{5A58BC92-EC25-49F9-991A-3A2F03E5899F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeHub.Portal", "src\frontend\CodeHub.Portal\CodeHub.Portal.csproj", "{E6AF518B-3FAC-4D3F-8BF2-7C8411B153CA}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Config", "Config", "{794E0C1D-6331-49C2-B835-D02C1FE83D01}"
@@ -36,7 +34,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Config", "Config", "{794E0C
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Config", "Config", "{21C8395D-3B75-46E3-AAAE-396CDE1679DB}"
 	ProjectSection(SolutionItems) = preProject
-		src\frontend\Directory.Build.props = src\frontend\Directory.Build.props
 		src\frontend\Frontend.Dockerfile = src\frontend\Frontend.Dockerfile
 		src\frontend\README.md = src\frontend\README.md
 	EndProjectSection
@@ -65,10 +62,6 @@ Global
 		{0B59B285-BE8E-48DA-9A9D-AFAF8B4D1C50}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0B59B285-BE8E-48DA-9A9D-AFAF8B4D1C50}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0B59B285-BE8E-48DA-9A9D-AFAF8B4D1C50}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5A58BC92-EC25-49F9-991A-3A2F03E5899F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A58BC92-EC25-49F9-991A-3A2F03E5899F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5A58BC92-EC25-49F9-991A-3A2F03E5899F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5A58BC92-EC25-49F9-991A-3A2F03E5899F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E6AF518B-3FAC-4D3F-8BF2-7C8411B153CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E6AF518B-3FAC-4D3F-8BF2-7C8411B153CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E6AF518B-3FAC-4D3F-8BF2-7C8411B153CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -103,7 +96,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{0B59B285-BE8E-48DA-9A9D-AFAF8B4D1C50} = {6BBAFA79-C6EC-4415-8249-3AF0BF69A8F8}
-		{5A58BC92-EC25-49F9-991A-3A2F03E5899F} = {FD7D4EDA-3EEC-4989-B479-563417E6848F}
 		{E6AF518B-3FAC-4D3F-8BF2-7C8411B153CA} = {FD7D4EDA-3EEC-4989-B479-563417E6848F}
 		{794E0C1D-6331-49C2-B835-D02C1FE83D01} = {6BBAFA79-C6EC-4415-8249-3AF0BF69A8F8}
 		{21C8395D-3B75-46E3-AAAE-396CDE1679DB} = {FD7D4EDA-3EEC-4989-B479-563417E6848F}
