+++ src/aspire/CodeHub.Aspire.AppHost/Program.cs
@@ -1,5 +1,3 @@
-using Aspire.Hosting;
-
 var builder = DistributedApplication.CreateBuilder(args);
 
 var api = builder.AddProject<Projects.CodeHub_Api>("api");
