+++ src/frontend/CodeHub.Portal/Features/Cloud/CloudHttpClient.cs
@@ -1,10 +1,8 @@
-using System.Net.Http.Json;
 using System.Text.Json;
 using CodeHub.Domain.Cloud;
 using CodeHub.Shared;
-using Microsoft.Extensions.Logging;
 
-namespace CodeHub.Portal.Services.Services;
+namespace CodeHub.Portal.Features.Cloud;
 
 public sealed class CloudHttpClient : ICloudHttpClient
 {
