+++ src/frontend/CodeHub.Portal/Features/Git/Client/GitHttpClient.cs
@@ -1,10 +1,8 @@
-﻿using System.Net.Http.Json;
-using System.Text.Json;
+﻿using System.Text.Json;
 using CodeHub.Domain.Git;
 using CodeHub.Shared;
-using Microsoft.Extensions.Logging;
 
-namespace CodeHub.Portal.Services.Services;
+namespace CodeHub.Portal.Features.Git.Client;
 
 public sealed class GitHttpClient : IGitHttpClient
 {
