+++ src/frontend/CodeHub.Portal/Features/Layout/NavMenu.razor
