+++ src/frontend/CodeHub.Portal/Features/Shared/PlatformIcons.cs
@@ -1,4 +1,4 @@
-﻿namespace CodeHub.Portal.Components;
+﻿namespace CodeHub.Portal.Features.Shared;
 
 public static class PlatformIcons
 {
