+++ src/frontend/CodeHub.Portal/Features/_Imports.razor
@@ -7,5 +7,4 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using CodeHub.Portal
-@using CodeHub.Portal.Components
 @using MudBlazor
