--- src/frontend/Directory.Build.props
@@ -1,18 +1,0 @@
-<Project>
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <LangVersion>latest</LangVersion>
-        <Features>strict</Features>
-        <WarningLevel>4</WarningLevel>
-        <Nullable>enable</Nullable>
-        <AnalysisMode>Default</AnalysisMode>
-        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-        <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-        <EnableRoslynAnalyzers>true</EnableRoslynAnalyzers>
-        <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
-        <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-        <RunAnalyzersDuringBuild>true</RunAnalyzersDuringBuild>
-        <RunAnalyzersDuringBuildOnly>true</RunAnalyzersDuringBuildOnly>
-    </PropertyGroup>
-</Project>
